Improved avatar appearance on Messages lists when using theme compat
Some WP themes accomplish responsiveness for images by using a max-width: 100%;
attribute for all images in the content area. (Twenty Twelve is an example.)
This declaration causes problems for the way that the Messages screen markup
is built, so that the container element for the avatar has no width, and thus
the avatar doesn't show up on Webkit browsers.

Fixes #4790

git-svn-id: https://buddypress.svn.wordpress.org/trunk@6839 cdf35c40-ae34-48e0-9cc9-0c9da1808c22