Use event delegation for AJAX buttons in Groups and Members directories
In r6635, BP's use of the deprecated jQuery.live() was switched to jQuery.on().
However, the change was not made in a way that preserved the behavior of live()
with respect to items added to the DOM dynamically, as per an AJAX callback.
We replicate that behavior by using jQuery.on() delegation on a persistent
parent element. That way, group directory buttons (Join/Leave) and member
directory buttons (Add Friend/Remove Friend) continue to work even when the
view is refreshed via AJAX, as in pagination or reordering.

Fixes #4931

git-svn-id: https://buddypress.svn.wordpress.org/trunk@6918 cdf35c40-ae34-48e0-9cc9-0c9da1808c22