Introduces BP_UnitTestCase::create_user()
Since BP requires users to have a last_activity date to be set for them to
appear in user queries, it'll almost always be the case that users created in
testcases will require a last_activity value. create_user() is a wrapper for
the WP factory method that sets this meta value.

git-svn-id: https://buddypress.svn.wordpress.org/trunk@6932 cdf35c40-ae34-48e0-9cc9-0c9da1808c22