Introduces bp_get_the_profile_field_errors_action()
This convenience function concatenates an action name for use in rendering the
errors associated with a given xprofile field during registration.

Previously, the action name was concatenated inline, using the _field_name_
value. This caused problems in the case of multiselect fields, which have a
field name appended with '[]' for other reasons. See #2173

Fixes #4978

git-svn-id: https://buddypress.svn.wordpress.org/trunk@7122 cdf35c40-ae34-48e0-9cc9-0c9da1808c22