Rework "Cancel Friendship Request" button UX in member directories
Member directories have friendship action buttons for each user: notably,
Add Friend for non-friends and Remove Friend for existing friends. When
A has sent a friendship request to B, the button becomes "Cancel Friendship
Request". However, it makes no sense for B to see the same button text for
the pending request, since it's not hers to cancel.

This changeset makes it so that B will see "Friendship Requested", and
the button will lead to B's Friend Requests page, so that she can accept
or reject the request using the latter existing UI.

Fixes #5157

Props terraling

git-svn-id: https://buddypress.svn.wordpress.org/trunk@7374 cdf35c40-ae34-48e0-9cc9-0c9da1808c22