Introduce BuddyPress Login Widget
This sidebar widget shows a set of login fields to logged-out users, along with
a link to the Register page (if registration is enabled). Logged-in users will
see their avatar, a link to their profile, and a Log Out link (as in the old
bp-default sidebar). This is particularly important now that BP can be used
with any theme, even themes that do not have obvious login mechanisms on the
front end.

Borrowed in part from bbPress, and in part from the BuddyPress Default theme.

Props modemlooper for an initial patch.

Fixes #1474

git-svn-id: https://buddypress.svn.wordpress.org/trunk@7450 cdf35c40-ae34-48e0-9cc9-0c9da1808c22