Clear activity item cache when it receives comments
r7997 fixed a bug was causing single activity items not to be cached properly
when a BP_Activity_Activity object was instantiated. This fix uncovered a bug
in activity commenting, where the activity parent was cached as childless, and
the cache was not invalidated when a new comment was posted.

See #5434

git-svn-id: https://buddypress.svn.wordpress.org/trunk@8004 cdf35c40-ae34-48e0-9cc9-0c9da1808c22