Templates: remove the easing plugin
The easing library we were using (in only two places) is a sizeable
chunk of code that would need a reasonable amount of modification to
bring up to WP standards.

Instead, I’ve decided to remove it and just use the ‘swing’ easing
type, which comes bundled with jQuery.

See #5613

git-svn-id: https://buddypress.svn.wordpress.org/trunk@8379 cdf35c40-ae34-48e0-9cc9-0c9da1808c22