Standardize placement of 'bp_get_activity_latest_update_excerpt' filter
When latest activity update excerpts are displayed in the Members directory,
they are displayed like this:

  - "Foo"

instead of just

  Foo

The wrapped string is then sent through the 'bp_get_activity_latest_update_excerpt'
filter. However, this creates inconsistencies with other uses of the
filter, which only filter the excerpt itself - not the dash and quotes. This
changeset fixes the inconsistency.

In addition, certain translations of the wrapper string were removing an
erroneous space character from the original string, which had been preventing
the make_clickable() callback from appending the trailing quote to URLs that
appeared at the end of the activity excerpt. In these translations (such as
fr_FR), the result was that the improper filter placement described above was
causing trailing quotation marks to be included as part of URLs when they
appeared at the end of an update string (from the point of view of make_clickable()).
This changeset also removes this extra space, since the bug will no longer
manifest itself in any language.

Fixes #5626

git-svn-id: https://buddypress.svn.wordpress.org/trunk@8400 cdf35c40-ae34-48e0-9cc9-0c9da1808c22