Specify logged-in user when getting first name for "What's new?" activity prompt
By default, BuddyPress only displays the post form in a context where
bp_get_user_fullname() would default to the logged-in user. However, if the
form is included elsewhere - such as on another user's profile - the first name
displayed would erroneously be that of the displayed user, rather than the
logged in user who is being prompted to post an update.

Fixes #5540

Props lenasterg

git-svn-id: https://buddypress.svn.wordpress.org/trunk@8405 cdf35c40-ae34-48e0-9cc9-0c9da1808c22