Add 'option' to the whitelist of xprofile field types
The 'option' field type is not full-fledged, in the sense that it has an
associated BP_XProfile_Field_Type class; it is used only to save the values
that are child options of type 'radio', 'select', 'multiselect' and other
option-supporting field types. It is important that 'option' thus be added
manually to the field type whitelist, so that option fields can be created
programatically.

Fixes #5472

git-svn-id: https://buddypress.svn.wordpress.org/trunk@8518 cdf35c40-ae34-48e0-9cc9-0c9da1808c22